+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- Range widget (absolute values) [#454](https://github.com/CartoDB/carto-react/pull/454)
+
 ## 1.4
 
 ### 1.4.0-alpha.0 (2022-07-14)